move Cell to inside GAT
