+++ Assets/Packages.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25bd4fcda2ae397489ab353693fcce0a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
