+++ Assets/Packages/SharpZipLib.1.3.1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d7435fd75d4b0c449179559017c8d14
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
