+++ Assets/Packages/SharpZipLib.1.3.1/lib.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 06d8fe32ba8c6a043960e902adf64b84
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
