+++ Assets/Packages/SharpZipLib.1.3.1/lib/net45.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8029ae3931ed03c4ca984d50731bed93
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
